+++ src/format.rs
@@ -13,7 +13,7 @@ pub struct Header {
     pub cols: u16,
     pub rows: u16,
     pub timestamp: u64,
-    pub idle_time_limit: Option<f32>,
+    pub idle_time_limit: Option<f64>,
     pub command: Option<String>,
     pub title: Option<String>,
     pub env: HashMap<String, String>,
